Moved the auction-related stuff to it's own package
